+++ api/profile/getLoggedProfileImage.php
@@ -18,14 +18,17 @@ if(isset($postdata) && !empty($postdata))
     }
     if($tokenValidity = Token::validate($authToken, $dbhash) == true){
       $payloadContent = Token::getPayload($authToken, $dbhash);
-      $payloadContent = $payloadContent['data'];
-      $response = [
-        'code' => 1,
-        'message' => 'Successful Authentication',
-        'tokenValidity' => $tokenValidity,
-        'tokenPayload' => $payloadContent,
-      ];
-      echo json_encode($response);
+      if($result = mysqli_query($con, $sql)){
+        while($row = mysqli_fetch_assoc($result)){
+          $rows[] = $row;
+          $db_image = $rows[0]['profile_image'];
+        }
+        $response = [
+          'code' => 1,
+          'image' => $db_image
+        ];
+        echo json_encode($response);
+      }
     }
     else{
       $response = [
@@ -45,5 +48,6 @@ else{
   ];
   echo json_encode($response);
 }
+
 
 ?>