+++ api/profile/getProfile.php
@@ -1,5 +1,0 @@
-<?php
-require '../database.php';
-use ReallySimpleJWT\Token;
-
-?>